# zmm/src/App/ClientApp/src/assets/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "MLW / ZMK",
        "description": "Swagger for MLW / ZMK",
        "version": "1.2.0"
    },
    "schemes": [
        "http",
        "https"
    ],
    "basePath": "/zmk/api/v1",
    "consumes": [
        
    ],
    "produces": [
        
    ],
    "securityDefinitions": {
        "Basic": {
            "type": "basic"
        }
    },
    "security": [
        {
            "Basic": []
        }
    ],
    "paths": {
...
        "/models": {
            "get": {
                "operationId": "models_list",
                "description": "Get the list of loaded models.",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Returns a list of models."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                },
                "tags": [
                    "Model"
                ]
            },
            "post": {
                "operationId": "models_post",
                "description": "Loads a model into memory.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                  ],
                "parameters": [
                    {
                        "name":"filePath",
                        "in":"formData",
                        "required":true,
                        "description":"Path of the file.",
                        "type":"string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a json response whether success or fail."
                    },
                    "400": {
                        "description": "Invalid Request Params."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                },
                "tags": [
                    "Model"
                ]
            },
            "parameters": []
        },
...
    },
...
}